requirements.txt
alembic==1.13.2
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
swagger_ui_bundle==1.1.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
